Improve how scope errors are handled to only fetch scopes when needed
